improved path handling and command string for final ffmpeg render

+++ dear_render.py
@@ -74,12 +74,12 @@ for fin in files:
 
     #p=os.popen('ffmpeg -y -r %d -sameq -i %simg%%05d.png -i %s %s'%(framerate,clean_outdir,clean_fin,clean_fout.rsplit('.',1)[0]+'_img.avi'))
     #p=os.popen('ffmpeg -y -framerate %d -i %sconv%%05d.png -i %s -vcodec libx264 -crf 18 -preset slow -vf "transpose=1" -r %d %s'%(framerate,clean_outdir,clean_fin,framerate,clean_fout))
-    p=os.popen('ffmpeg -y -framerate %d -i %sconv%%05d.png -i %s -vcodec libx264 -pix_fmt yuv420p -acodec copy -b:v 8000k -f mov -vf "transpose=1" -r %d %s'%(framerate,clean_outdir,clean_fin,framerate,clean_fout))
+    p=os.popen('ffmpeg -y -framerate %d -i "%sconv%%05d.png" -i "%s" -vcodec libx264 -pix_fmt yuv420p -acodec copy -b:v 8000k -f mov -vf "transpose=1" -r %d "%s"'%(framerate,clean_outdir,clean_fin,framerate,clean_fout))
     p.close()
     print '\n'
     print 'rendered video'
 
-    if not args.keep_frames:
+    if not args.keep_frames and os.path.isfile(clean_fout):
         shutil.rmtree(outdir)
         os.remove(wav)
         print 'cleaned temp stuff'
